+++ frontend/src/pages/Home.tsx
@@ -1,4 +1,5 @@
-import Header from "@/components/Header";
+
+import Header from "@/components/Header/Header";
 import { UserButton } from "@clerk/clerk-react";
